+++ src/Lang/Parser.hs
@@ -1,17 +1,18 @@
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Lang.Parser where
 
 import           Control.Monad
-import           Import                  hiding ( optional
+import           Import                  hiding ( many
+                                                , optional
                                                 , try
                                                 , (<|>)
                                                 )
 import           Lang.Lexer
 import           Text.Parsec
-import           Text.ParserCombinators.Parsec
-                                         hiding ( try )
+import           Text.Parsec.String
 import           Text.ParserCombinators.Parsec.Expr
 
 parseProg :: String -> Either String Prog
@@ -19,8 +20,13 @@ parseProg p = case parse progParser "PiG" p of
   Left  err  -> Left $ show err
   Right prog -> Right prog
 
+parseFile :: FilePath -> IO (Either String [Prog])
+parseFile f = parseFromFile (sepEndBy1 progParser semi <* eof) f >>= \case
+  Left  err  -> return . Left $ show err
+  Right prog -> return $ Right prog
+
 progParser :: Parser Prog
-progParser = whiteSpace >> stmtParser
+progParser = whiteSpace >> (Stmt <$> stmtParser <|> Drct <$> drctParser)
 
 endParser :: ParsecT String u Identity ()
 endParser =
@@ -37,9 +43,22 @@ endParser =
        <|> eof
        )
 
+last :: Parser a -> Parser a
+last b = b <* endParser
+
+drctParser :: Parser Drct
+drctParser =
+  try (last $ reserved ":clear" >> return Clear)
+    <|> try (last $ reserved ":exit" >> return Exit)
+    <|> try (last $ reserved ":help" >> return Help)
+    <|> Rm
+    <$> try (last $ reserved ":rm" >> identifier)
+    <|> Load
+    <$> try (last $ reserved ":load" >> stringLiteral)
+
 stmtParser :: Parser Stmt
 stmtParser = do
-  list <- (sepEndBy1 singleStmtParser semi)
+  list <- (sepEndBy1 singleStmtParser (semi <|> many1 endOfLine))
   return $ case list of
     [stmt] -> stmt
     _      -> Seq list
@@ -77,7 +96,7 @@ assignStmtParser =
   (do
       var  <- identifier
       expr <- reservedOp "=" >> exprParser
-      return $ var := expr
+      return $ Assign var expr
     )
     <?> "assignment"
 
@@ -109,7 +128,7 @@ funValParser =
       try
           (braces $ do
             body <- stmtParser
-            ret  <- reserved "return" >> exprParser
+            ret  <- reserved "return" >> exprParser <* semi
             return $ FunVal args body ret
           )
         <|> (do
@@ -131,7 +150,6 @@ exprParser =
     <|> try (last algExprParser)
     <|> try (last funExprParser)
     <|> last listExprParser
-  where last b = b <* endParser
 
 algExprParser :: Parser Expr
 algExprParser = buildExpressionParser algOperators algTerm